+++ apps/backend/src/app/api/v1/users/crud.tsx
@@ -23,33 +23,14 @@ export const userFullInclude = {
       providerConfig: true,
     },
   },
-  contactChannels: true,
   authMethods: {
     include: {
       passwordAuthMethod: true,
       otpAuthMethod: true,
-      oauthAuthMethod: {
-        include: {
-          oauthProviderConfig: {
-            include: {
-              proxiedOAuthConfig: true,
-              standardOAuthConfig: true,
-            },
-          }
-        }
-      },
+      oauthAuthMethod: true,
     }
   },
-  connectedAccounts: {
-    include: {
-      oauthProviderConfig: {
-        include: {
-          proxiedOAuthConfig: true,
-          standardOAuthConfig: true,
-        },
-      }
-    }
-  },
+  contactChannels: true,
   teamMembers: {
     include: {
       team: true,
@@ -90,8 +71,6 @@ export const userPrismaToCrud = (
     throw new StackAssertionError("User cannot have more than one selected team; this should never happen");
   }
 
-  const contactChannels = prisma.contactChannels.map(c => contactChannelToCrud(c));
-
   // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
   const primaryEmailContactChannel = prisma.contactChannels.find((c) => c.type === 'EMAIL' && c.isPrimary);
   const passwordAuth = prisma.authMethods.find((m) => m.passwordAuthMethod);
@@ -118,7 +97,6 @@ export const userPrismaToCrud = (
     selected_team_id: selectedTeamMembers[0]?.teamId ?? null,
     selected_team: selectedTeamMembers[0] ? teamPrismaToCrud(selectedTeamMembers[0]?.team) : null,
     last_active_at_millis: lastActiveAtMillis,
-    contact_channels: contactChannels,
   };
 };
 
@@ -471,10 +449,8 @@ export const usersCrudHandlers = createLazyProxy(() => createCrudHandlers(usersC
             `${data.display_name}'s Team` :
             data.primary_email ?
               `${data.primary_email}'s Team` :
-              "Personal Team"
-        },
-        query: {
-          add_current_user: "true",
+              "Personal Team",
+          creator_user_id: 'me',
         },
         project: auth.project,
         user: result,