+++ apps/e2e/tests/backend/endpoints/api/v1/auth/password/sign-in.test.ts
@@ -25,16 +25,6 @@ it("should allow signing in to existing accounts", async ({ expect }) => {
         "auth_with_email": true,
         "client_metadata": null,
         "client_read_only_metadata": null,
-        "contact_channels": [
-          {
-            "id": "<stripped UUID>",
-            "is_primary": true,
-            "is_verified": false,
-            "type": "email",
-            "used_for_auth": true,
-            "value": "<stripped UUID>@stack-generated.example.com",
-          },
-        ],
         "display_name": null,
         "has_password": true,
         "id": "<stripped UUID>",
@@ -51,6 +41,7 @@ it("should allow signing in to existing accounts", async ({ expect }) => {
     }
   `);
 });
+// TODO: check auth methods
 
 it("should not allow signing in with an e-mail that never signed up", async ({ expect }) => {
   const response = await niceBackendFetch("/api/v1/auth/password/sign-in", {