+++ apps/e2e/tests/backend/endpoints/api/v1/auth/password/sign-up.test.ts
@@ -34,16 +34,6 @@ it("should sign up new users", async ({ expect }) => {
         "auth_with_email": true,
         "client_metadata": null,
         "client_read_only_metadata": null,
-        "contact_channels": [
-          {
-            "id": "<stripped UUID>",
-            "is_primary": true,
-            "is_verified": false,
-            "type": "email",
-            "used_for_auth": true,
-            "value": "<stripped UUID>@stack-generated.example.com",
-          },
-        ],
         "display_name": null,
         "has_password": true,
         "id": "<stripped UUID>",