+++ packages/stack-shared/src/interface/crud/users.ts
@@ -40,7 +40,6 @@ export const usersCrudServerReadSchema = fieldSchema.yupObject({
   client_read_only_metadata: fieldSchema.userClientReadOnlyMetadataSchema,
   server_metadata: fieldSchema.userServerMetadataSchema,
   last_active_at_millis: fieldSchema.userLastActiveAtMillisSchema.required(),
-  contact_channels: fieldSchema.yupArray(contactChannelSchema).required(),
 
   oauth_providers: fieldSchema.yupArray(fieldSchema.yupObject({
     id: fieldSchema.yupString().required(),