+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stackframe/stack-backend
 
+## 2.6.1
+
+### Patch Changes
+
+- Bugfixes
+  - @stackframe/stack-emails@2.6.1
+  - @stackframe/stack-shared@2.6.1
+
 ## 2.6.0
 
 ### Minor Changes