+++ examples/express-proxied-server/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @stackframe/proxied-server-example
 
+## 2.6.1
+
 ## 2.6.0
 
 ### Minor Changes